fix: 정속주행률 데이터타입 변경

+++ src/main/java/com/smooth/driving_analysis_service/reports/basic_summary/dto/BasicSummaryResponseDto.java
@@ -30,5 +30,5 @@ public class BasicSummaryResponseDto {
 
     private Double averageSpeedKmh;           //평균 속도
     
-    private Double averageCruiseRatio;        //평균 정속중행률
+    private int averageCruiseRatio;        //평균 정속중행률
 }
